requirements.txt
Flask==2.3.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0